RavenDB-19561 - Addressed PR Comments
