+++ test/Tests.Infrastructure/RavenTestBase.cs
@@ -8,6 +8,7 @@ using System.Net.Sockets;
 using System.Runtime.CompilerServices;
 using System.Runtime.InteropServices;
 using System.Security.Cryptography.X509Certificates;
+using System.Text;
 using System.Threading;
 using System.Threading.Tasks;
 using Raven.Client;
@@ -791,6 +792,25 @@ namespace FastTests
             tcpListener.Stop();
 
             return port;
+        }
+
+        public static string GenRandomString(int size)
+        {
+            return GenRandomString(new Random(), size);
+        }
+
+        public static string GenRandomString(Random random, int size)
+        {
+            var sb = new StringBuilder(size);
+            // var ran = new Random();
+            var firstCharAsInt = Convert.ToInt32('a');
+            var lastCharAsInt = Convert.ToInt32('z');
+            for (int i = 0; i < size; i++)
+            {
+                sb.Append(Convert.ToChar(random.Next(firstCharAsInt, lastCharAsInt + 1)));
+            }
+
+            return sb.ToString();
         }
     }
 }
